Split functions in different files
